requirements.txt
aiomysql==0.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
rpds-py==0.10.4
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
mysqlclient
requests~=2.31.0
geopy~=2.4.0
shapely~=2.0.2
schedule~=1.2.1